Move GridDetails component to own component file
