+++ src/client/components/GridDetails/GridDetails.tsx
